--- src/pymmcore_widgets/hcs/__init__.py
@@ -1,1 +1,0 @@
-"""High content screening widgets for PyMMCore."""
