+++ src/pymmcore_widgets/useq_widgets/points_plans/__init__.py
@@ -2,6 +2,12 @@
 
 from ._grid_row_column_widget import GridRowColumnWidget
 from ._points_plan_selector import RelativePointPlanSelector
+from ._points_plan_widget import PointsPlanWidget
 from ._random_points_widget import RandomPointWidget
 
-__all__ = ["GridRowColumnWidget", "RandomPointWidget", "RelativePointPlanSelector"]
+__all__ = [
+    "GridRowColumnWidget",
+    "RandomPointWidget",
+    "RelativePointPlanSelector",
+    "PointsPlanWidget",
+]
