+++ src/pymmcore_widgets/useq_widgets/points_plans/_points_plan_selector.py
@@ -58,6 +58,9 @@ class _FovWidget(QDoubleSpinBox):
 class RelativePointPlanSelector(QWidget):
     """Widget to select a relative multi-position point plan.
 
+    See also: [PointsPlanWidget][pymmcore_widgets.useq_widgets.PointsPlanWidget]
+    which combines this widget with a graphical representation of the points.
+
     In useq, a RelativeMultiPointPlan is one of:
     - useq.RelativePosition
     - useq.RandomPoints
